Move collision scenes out of Services namespace
