+++ src/Detach.Demos.Collisions/CollisionScenes/LineOrientedRectangle.cs
@@ -4,15 +4,15 @@ using Detach.Demos.Collisions.Utils;
 using ImGuiNET;
 using System.Numerics;
 
-namespace Detach.Demos.Collisions.Services.CollisionScenes;
+namespace Detach.Demos.Collisions.CollisionScenes;
 
-public sealed class LineLine : CollisionScene<LineSegment2D, LineSegment2D>
+public sealed class LineOrientedRectangle : CollisionScene<LineSegment2D, OrientedRectangle>
 {
 	private const float _linePointOffsetA = 64;
-	private const float _linePointOffsetB = 128;
+	private const float _rectangleOffset = 128;
 
-	public LineLine()
-		: base(Geometry2D.LineLine)
+	public LineOrientedRectangle()
+		: base(Geometry2D.LineOrientedRectangle)
 	{
 	}
 
@@ -25,9 +25,10 @@ public sealed class LineLine : CollisionScene<LineSegment2D, LineSegment2D>
 		A = new LineSegment2D(
 			CollisionSceneConstants.Origin + new Vector2(MathF.Cos(quarterTime) * _linePointOffsetA, MathF.Sin(quarterTime) * _linePointOffsetA),
 			CollisionSceneConstants.Origin + new Vector2(MathF.Cos(TotalTime) * _linePointOffsetA, MathF.Sin(halfTime) * _linePointOffsetA));
-		B = new LineSegment2D(
-			CollisionSceneConstants.Origin + new Vector2(MathF.Cos(halfTime) * _linePointOffsetB, MathF.Sin(halfTime) * _linePointOffsetB),
-			CollisionSceneConstants.Origin + new Vector2(MathF.Cos(quarterTime) * _linePointOffsetB, MathF.Sin(quarterTime) * _linePointOffsetB));
+		B = new OrientedRectangle(
+			CollisionSceneConstants.Origin + new Vector2(MathF.Cos(TotalTime) * _rectangleOffset, MathF.Sin(TotalTime) * _rectangleOffset),
+			new Vector2(64 + MathF.Sin(TotalTime) * 32, 32 + MathF.Cos(TotalTime) * 16),
+			TotalTime * 1.5f);
 	}
 
 	public override void Render()
@@ -35,6 +36,6 @@ public sealed class LineLine : CollisionScene<LineSegment2D, LineSegment2D>
 		PositionedDrawList drawList = new(ImGui.GetWindowDrawList(), ImGui.GetCursorScreenPos());
 		drawList.AddBackground(CollisionSceneConstants.Size);
 		drawList.AddLine(A, HasCollision);
-		drawList.AddLine(B, HasCollision);
+		drawList.AddOrientedRectangle(B, HasCollision);
 	}
 }
