+++ src/Detach.Demos.Collisions/CollisionScenes/PointInOrientedRectangle.cs
@@ -4,7 +4,7 @@ using Detach.Demos.Collisions.Utils;
 using ImGuiNET;
 using System.Numerics;
 
-namespace Detach.Demos.Collisions.Services.CollisionScenes;
+namespace Detach.Demos.Collisions.CollisionScenes;
 
 public sealed class PointInOrientedRectangle : CollisionScene<Vector2, OrientedRectangle>
 {
