+++ src/Detach.Demos.Collisions/CollisionScenes/PointInRectangle.cs
@@ -4,15 +4,15 @@ using Detach.Demos.Collisions.Utils;
 using ImGuiNET;
 using System.Numerics;
 
-namespace Detach.Demos.Collisions.Services.CollisionScenes;
+namespace Detach.Demos.Collisions.CollisionScenes;
 
-public sealed class LineRectangle : CollisionScene<LineSegment2D, Rectangle>
+public sealed class PointInRectangle : CollisionScene<Vector2, Rectangle>
 {
-	private const float _linePointOffsetA = 64;
+	private const float _pointOffset = 64;
 	private const float _rectangleOffset = 128;
 
-	public LineRectangle()
-		: base(Geometry2D.LineRectangle)
+	public PointInRectangle()
+		: base(Geometry2D.PointInRectangle)
 	{
 	}
 
@@ -20,11 +20,8 @@ public sealed class LineRectangle : CollisionScene<LineSegment2D, Rectangle>
 	{
 		base.Update(dt);
 
-		float halfTime = TotalTime / 2;
-		float quarterTime = TotalTime / 4;
-		A = new LineSegment2D(
-			CollisionSceneConstants.Origin + new Vector2(MathF.Cos(quarterTime) * _linePointOffsetA, MathF.Sin(quarterTime) * _linePointOffsetA),
-			CollisionSceneConstants.Origin + new Vector2(MathF.Cos(TotalTime) * _linePointOffsetA, MathF.Sin(halfTime) * _linePointOffsetA));
+		float doubleTime = TotalTime * 2;
+		A = CollisionSceneConstants.Origin + new Vector2(MathF.Cos(doubleTime) * _pointOffset, MathF.Sin(doubleTime) * _pointOffset);
 		B = Rectangle.FromCenter(
 			CollisionSceneConstants.Origin + new Vector2(MathF.Cos(TotalTime) * _rectangleOffset, MathF.Sin(TotalTime) * _rectangleOffset),
 			new Vector2(160 + MathF.Sin(TotalTime) * 32));
@@ -34,7 +31,7 @@ public sealed class LineRectangle : CollisionScene<LineSegment2D, Rectangle>
 	{
 		PositionedDrawList drawList = new(ImGui.GetWindowDrawList(), ImGui.GetCursorScreenPos());
 		drawList.AddBackground(CollisionSceneConstants.Size);
-		drawList.AddLine(A, HasCollision);
+		drawList.AddPoint(A, HasCollision);
 		drawList.AddRectangle(B, HasCollision);
 	}
 }
